+++ cmd/hidden_lake/helpers/encryptor/internal/handler/index_test.go
@@ -25,7 +25,7 @@ func TestErrorsAPI(t *testing.T) {
 		),
 	)
 
-	if _, err := client.EncryptMessage(context.Background(), asymmetric.LoadRSAPubKey(testutils.TgPubKeys[0]), []byte{123}); err == nil {
+	if _, err := client.EncryptMessage(context.Background(), asymmetric.LoadRSAPubKey(testutils.TgPubKeys[0]), payload.NewPayload(1, []byte{123})); err == nil {
 		t.Error("success encrypt message with unknown host")
 		return
 	}