+++ cmd/hidden_lake/helpers/encryptor/pkg/settings/types.go
@@ -2,5 +2,6 @@ package settings
 
 type SContainer struct {
 	FPublicKey string `json:"public_key"`
+	FPldHead   uint64 `json:"pld_head"`
 	FHexData   string `json:"hex_data"`
 }